Extract git-backed build logic into utils module
